refactor: DAO를 인터페이스 - 구현체 형태에서 구현체 형태로 변경
